Conexão com Banco de Dados Sql server, Testes e correção de erros

+++ BackEnd/apiDespesasPessoais/Services/IUsuarioService.cs
@@ -6,9 +6,9 @@ namespace apiDespesasPessoais.Services
     public interface IUsuarioService
     {
         Usuario Create(Usuario usuario);
-        Usuario FindById(long idUsuario);
+        Usuario FindById(int idUsuario);
         List<Usuario> FindAll();
         Usuario Update(Usuario usuario);
-        void Delete(long id);
+        void Delete(int id);
     }
 }
